通过测试各个功能修复若干bug；增：文件的下载。基础功能均可使用

+++ src/main/java/cn/fts/plugin/timerTask/FileCheckTask.java
@@ -2,7 +2,15 @@ package cn.fts.plugin.timerTask;
 
 import cn.fts.po.File;
 import cn.fts.service.FileService;
+import cn.fts.utils.PageUtil;
+import cn.fts.utils.SpringContextUtil;
+import cn.fts.utils.TimeUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+import org.springframework.context.support.FileSystemXmlApplicationContext;
+import org.springframework.stereotype.Component;
+import org.springframework.web.context.support.XmlWebApplicationContext;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -15,20 +23,29 @@ import java.util.TimerTask;
  *  当前时间，比对是否超期，如果超期，将其删除，发送通知给对应用户，记录日志。
  */
 
+@Component
 public class FileCheckTask extends TimerTask {
     private static boolean isRunning = false;
+    private int checkCount = 1;
 
-    @Autowired
     FileService fileService;
+
+    public FileCheckTask() {
+        fileService = SpringContextUtil.getBean("fileService");
+    }
 //    @Autowired
 //    NoticeService noticeService
 
     public void run() {
         if (!isRunning) {
             isRunning = true;
+            checkCount++;
             List<File> fileList = fileService.select();
             List<String> overdueIdList = checkOverdueFile(fileList);
-            fileService.deleteBatchByPrimaryKey(overdueIdList);
+            if (overdueIdList != null && overdueIdList.size() > 0) {
+                fileService.deleteBatchByPrimaryKey(overdueIdList);
+            }
+            System.out.println(TimeUtils.dateToString() + " 第 " + checkCount + " 次检查，被删除的文件ID：\t" + overdueIdList);
             isRunning = false;
         } else {
         }
@@ -40,8 +57,8 @@ public class FileCheckTask extends TimerTask {
         for (File file:
                 fileList) {
             Date start = file.getStart();
-            Date keep = new Date(file.getKeep()*60*1000);
-            if ((start.getTime()+keep.getTime()) > now.getTime()) {
+            if ((start.getTime()+file.getKeep()*60*1000) < now.getTime()) {
+//                System.out.println("overDue file id :\t" + file.getFileid() + "\tstart:\t" + start.getTime() + "\tkeep:\t" + file.getKeep() + "\tstart + keep:\t" + (start.getTime()+file.getKeep()*60*1000) + "\tnow:\t" + now.getTime());
                 overdueIdList.add(file.getFileid());
             }
         }
